Implemented makeRAW, makeCanonical, readLine, setBaudrate and isRAWTerminal methods.
Now this update makes use of the RAW and Canonical transmission
methods, witch makes it faster and more reliable. The readUntilChar
method was modified and become readLine method. Now the terminal
underneath reads a line and returns it almost ready. The method
converts ‘\n’ or ‘\r’, witch ever it finds first, to a NULL character,
this method is used on Canonical mode.
The makeRAW and makeCanonical methods allow the user to change the
transmission mode on the fly.
The isRAWTerminal method allow the user to know if the terminal is
working on a RAW, or Canonical way, or if it was set at all.
Now the baudrate is set on a separate method, the setBaudrate, and not
when you open the port.